requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
loguru==0.7.2

sqlalchemy==2.0.32
alembic==1.13.2

pillow==10.4.0
numpy==2.0.2

scikit-learn==1.5.1
pandas==2.2.2
tqdm==4.66.5
matplotlib==3.9.2
opencv-python==4.10.0.84
pyyaml==6.0.2

torch==2.3.0
torchvision==0.18.0